+++ org.mixedrealitytoolkit.uxcomponents/CHANGELOG.md
@@ -7,3 +7,4 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 ### Fixed
 
 * Fixed support for UPM package publishing in the Unity Asset Store.
+* Fixed Bug 643 - Experimental SimpleEmptyButton and SimpleActionButton prefabs have missing "See It Say It Label" GameObject reference